requirements/local.txt
asgiref==3.4.1
Django==3.2.7
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-model-utils==4.1.1
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2==2.9.1
pytz==2019.3
sqlparse==0.3.1
Unipath==1.1